AAAAAAPOOOP
Okay, so Now I added the 3D graphical aspect of the game into the
client, so if you run it, and it successfully connects, it will actually
display the game, and work in terms of movement. It does not handle
retrieval of information nor sending yet. I also changed up the server
to avoid some errors.